+++ examples/common/asynchronous_server.py
@@ -108,7 +108,7 @@ def run_async_server():
     identity.VendorUrl = 'http://github.com/bashwork/pymodbus/'
     identity.ProductName = 'Pymodbus Server'
     identity.ModelName = 'Pymodbus Server'
-    identity.MajorMinorRevision = '2.2.0'
+    identity.MajorMinorRevision = '2.3.0'
     
     # ----------------------------------------------------------------------- # 
     # run the server you want