+++ examples/common/dbstore_update_server.py
@@ -86,7 +86,7 @@ def run_dbstore_update_server():
     identity.VendorUrl = 'http://github.com/bashwork/pymodbus/'
     identity.ProductName = 'pymodbus Server'
     identity.ModelName = 'pymodbus Server'
-    identity.MajorMinorRevision = '2.2.0'
+    identity.MajorMinorRevision = '2.3.0'
 
     # ----------------------------------------------------------------------- #
     # run the server you want